backend/requirements.txt
fastapi
uvicorn
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
pydantic>=2.0.0
python-dotenv
langgraph
firebase-admin
